make public the stuff
