+++ tests/oldcheck/oldcheck_test.go
@@ -1,0 +1,43 @@
+package oldcheck
+
+import (
+	"testing"
+
+	"github.com/openfga/openfga/cmd/run"
+	"github.com/openfga/openfga/tests"
+	"github.com/stretchr/testify/require"
+	pb "go.buf.build/openfga/go/openfga/api/openfga/v1"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
+)
+
+func TestCheckMemory(t *testing.T) {
+	testCheck(t, "memory")
+}
+
+func TestCheckPostgres(t *testing.T) {
+	testCheck(t, "postgres")
+}
+
+func TestCheckMySQL(t *testing.T) {
+	testCheck(t, "mysql")
+}
+
+func testCheck(t *testing.T, engine string) {
+
+	cfg := run.MustDefaultConfigWithRandomPorts()
+	cfg.Log.Level = "none"
+	cfg.Datastore.Engine = engine
+
+	cancel := tests.StartServer(t, cfg)
+	defer cancel()
+
+	conn, err := grpc.Dial(cfg.GRPC.Addr,
+		grpc.WithBlock(),
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+	)
+	require.NoError(t, err)
+	defer conn.Close()
+
+	RunTests(t, pb.NewOpenFGAServiceClient(conn))
+}
